extract out manual_group_on_time_aggregation func
