+++ src/sentry/search/events/builder/issue_platform.py
@@ -1,7 +1,6 @@
 from typing import List, Optional
 
-from snuba_sdk import Column, Function
-
+from sentry.issues.query import manual_group_on_time_aggregation
 from sentry.search.events.builder import TimeseriesQueryBuilder
 from sentry.search.events.types import ParamsType
 from sentry.utils.snuba import Dataset
@@ -34,39 +33,5 @@ class IssuePlatformTimeseriesQueryBuilder(TimeseriesQueryBuilder):
             has_metrics=has_metrics,
             skip_tag_resolution=skip_tag_resolution,
         )
-        rollup_to_start_func = {
-            60: "toStartOfMinute",
-            3600: "toStartOfHour",
-            3600 * 24: "toDate",
-        }
-        rollup_func = rollup_to_start_func.get(interval)
-        if rollup_func:
-            if rollup_func == "toDate":
-                self.time_column = Function(
-                    "toUnixTimestamp",
-                    [Function("toDateTime", [Function(rollup_func, [Column("timestamp")])])],
-                    alias="time",
-                )
-            else:
-                self.time_column = Function(
-                    "toUnixTimestamp", [Function(rollup_func, [Column("timestamp")])], alias="time"
-                )
-        else:
-            self.time_column = Function(
-                "multiply",
-                [
-                    Function(
-                        "intDiv",
-                        [
-                            Function(
-                                "toUInt32", [Function("toUnixTimestamp", [Column("timestamp")])]
-                            ),
-                            interval,
-                        ],
-                    ),
-                    interval,
-                ],
-                alias="time",
-            )
-
+        self.time_column = manual_group_on_time_aggregation(interval, "time")
         self.groupby = [self.time_column]
