+++ src/sentry/tsdb/snuba.py
@@ -24,6 +24,7 @@ from snuba_sdk.query import SelectableExpression
 
 from sentry.constants import DataCategory
 from sentry.ingest.inbound_filters import FILTER_STAT_KEYS_TO_VALUES
+from sentry.issues.query import manual_group_on_time_aggregation
 from sentry.tsdb.base import BaseTSDB, TSDBModel
 from sentry.utils import outcomes, snuba
 from sentry.utils.dates import to_datetime
@@ -267,48 +268,6 @@ class SnubaTSDB(BaseTSDB):
 
         return known_rollups if known_rollups else synthetic_rollup
 
-    def __manual_group_on_time_aggregation_sqnl(
-        self, rollup, time_column_alias
-    ) -> SelectableExpression:
-        def rollup_agg(rollup_granularity: int, alias: str):
-            if rollup_granularity == 60:
-                return Function(
-                    "toUnixTimestamp", [Function("toStartOfMinute", [Column("timestamp")])], alias
-                )
-            elif rollup_granularity == 3600:
-                return Function(
-                    "toUnixTimestamp", [Function("toStartOfHour", [Column("timestamp")])], alias
-                )
-            elif rollup_granularity == 3600 * 24:
-                return Function(
-                    "toUnixTimestamp",
-                    [Function("toDateTime", [Function("toDate", [Column("timestamp")])])],
-                    alias,
-                )
-            else:
-                return None
-
-        # if we don't have an explicit function mapped to this rollup, we have to calculate it on the fly
-        # multiply(intDiv(toUInt32(toUnixTimestamp(timestamp)), granularity)))
-        synthetic_rollup = Function(
-            "multiply",
-            [
-                Function(
-                    "intDiv",
-                    [
-                        Function("toUInt32", [Function("toUnixTimestamp", [Column("timestamp")])]),
-                        rollup,
-                    ],
-                ),
-                rollup,
-            ],
-            time_column_alias,
-        )
-
-        known_rollups = rollup_agg(rollup, time_column_alias)
-
-        return known_rollups if known_rollups else synthetic_rollup
-
     def get_data(
         self,
         model,
@@ -427,7 +386,7 @@ class SnubaTSDB(BaseTSDB):
         ]
 
         if group_on_time and manual_group_on_time:
-            aggregations.append(self.__manual_group_on_time_aggregation_sqnl(rollup, "time"))
+            aggregations.append(manual_group_on_time_aggregation(rollup, "time"))
 
         if keys:
             start = to_datetime(series[0])
